content/media/webaudio/test/test_AudioContext.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  ok(!window.mozAudioContext, "mozAudioContext should be hidden behind a pref");
  var accessThrows = false;
  try {
    new mozAudioContext();
  } catch (e) {
    accessThrows = true;
  }
  ok(accessThrows, "mozAudioContext should be hidden behind a pref");
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);
  var ac = new mozAudioContext();
  ok(ac, "Create a mozAudioContext object");
  SpecialPowers.clearUserPref("media.webaudio.enabled");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
